docs: jsファイルの位置を変更
