+++ app/javascript/packs/application.js
@@ -18,9 +18,9 @@ import "./diagnosis/new";
 import "./shared/description";
 import "./shared/hamburger_menu";
 import "./shared/progress_bar";
-import "./shared/diagnosis_progress_bar";
+import "./diagnosis/diagnosis_progress_bar";
 import "./architecture/preview";
+import "./architecture/keep_search_params";
+import "./architecture/tag";
 import "./registrations/edit";
 import "./users/show";
-import "./shared/keep_search_params";
-import "./shared/tag";
