+++ app/javascript/packs/architecture/keep_search_params.js
