+++ app/javascript/packs/diagnosis/diagnosis_progress_bar.js
